Convert to Maven-project
Adding pom.xml and restructure project
